% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_of_stream.R
\name{analysis_of_stream}
\alias{analysis_of_stream}
\title{Generates full analysis of one stream of streamer}
\usage{
analysis_of_stream(path, window_size = 5, interactive = TRUE)
}
\value{
list of plots
}
\description{
Generates full analysis of one stream of streamer
}
