% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_of_streamer.R
\name{analysis_of_streamer}
\alias{analysis_of_streamer}
\title{Generates full analysis of all streams of streamer}
\usage{
analysis_of_streamer(path, window_size = 5, interactive = TRUE)
}
\value{
plots
}
\description{
Generates full analysis of all streams of streamer
}
