% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_for_row_spill_multi.R
\name{correct_for_row_spill_multi}
\alias{correct_for_row_spill_multi}
\title{Corrects for when .html to .txt spills over single comment to multiple lines.
This seems to happen when the comment has a specific emoji (not super clear)}
\usage{
correct_for_row_spill_multi(df)
}
\value{
df
}
\description{
Corrects for when .html to .txt spills over single comment to multiple lines.
This seems to happen when the comment has a specific emoji (not super clear)
}
