% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comments_over_time.R
\name{plot_comments_over_time}
\alias{plot_comments_over_time}
\title{Plots moving windowed comment of some moment over some stream}
\usage{
plot_comments_over_time(
  df,
  stream_title,
  window_size = 5,
  .f = mean,
  .fname = "Mean",
  interactive = FALSE
)
}
\arguments{
\item{.f}{function of moment}

\item{.fname}{name of function of moment for ggplot title}

\item{interactive}{}
}
\value{
ggplot (time series of some moment of twitch comments)
}
\description{
Plots moving windowed comment of some moment over some stream
}
