% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_comment_file.R
\name{process_comment_file}
\alias{process_comment_file}
\title{Processes .txt files of comments}
\usage{
process_comment_file(address)
}
\arguments{
\item{address}{Address of translated twitch comments .txt}
}
\value{
df
}
\description{
Processes .txt files of comments
}
