% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_window.R
\name{set_window}
\alias{set_window}
\title{Sets moving window average
TODO expand to w/e function not just moving ave (or have a separate func?)}
\usage{
set_window(df, window_size, .f = sum)
}
\arguments{
\item{window_size}{}
}
\value{

}
\description{
Sets moving window average
TODO expand to w/e function not just moving ave (or have a separate func?)
}
